+++ woodwork/tests/utils/test_accessor_utils.py
@@ -1,3 +1,4 @@
+import numpy as np
 import pandas as pd
 import pytest
 
@@ -43,6 +44,21 @@ def test_init_series_with_invalid_type(sample_df):
         error_message = f'Input must be of series type. The current input is of type {type(input_)}'
         with pytest.raises(TypeError, match=error_message):
             init_series(input_)
+
+
+def test_init_series_with_np_array(sample_series_pandas):
+    series = init_series(np.array(['a', 'b', 'c', 'a']))
+    series2 = init_series(sample_series_pandas)  # Sample series panda contains ['a','b','c','a']
+    assert series.equals(series2)
+    assert series.ww.logical_type == series2.ww.logical_type
+    assert series.ww.semantic_tags == series2.ww.semantic_tags
+
+
+def test_init_series_with_multidimensional_np_array():
+    input_ = np.array([['a', 'b'], ['a', 'b']])
+    error_message = f'np.ndarray input must be 1 dimensional. Current np.ndarray is {input_.ndim} dimensional'
+    with pytest.raises(ValueError, match=error_message):
+        init_series(input_)
 
 
 def test_init_series_valid_conversion_inferred_ltype(sample_series):