+++ woodwork/accessor_utils.py
@@ -34,7 +34,7 @@ def init_series(series, logical_type=None, semantic_tags=None,
     if not _is_series(series):
         if isinstance(series, (np.ndarray, pd.api.extensions.ExtensionArray)) and series.ndim == 1:
             series = pd.Series(series)
-        elif isinstance(series, (np.ndarray, pd.api.extensions.ExtensionArray)) and series.ndim != 1:
+        elif isinstance(series, np.ndarray) and series.ndim != 1:
             raise ValueError(f'np.ndarray input must be 1 dimensional. Current np.ndarray is {series.ndim} dimensional')
         else:
             raise TypeError(f'Input must be of series type. The current input is of type {type(series)}')