+++ docs/source/release_notes.rst
@@ -11,6 +11,7 @@ Future Release
     * Changes
         * Remove lines in ``LogicalType.transform`` that raise error if dtype conflicts (:pr:`1012`)
         * Add ``infer_datetime_format`` param to speed up ``to_datetime`` calls (:pr:`1016`)
+        * The default logical type is now the ``Unknown`` type instead of the ``NaturalLanguage`` type (:pr:`992`)
         * Add support for numpy inputs to Woodwork (:pr:`1023`)
     * Documentation Changes
     * Testing Changes
@@ -18,6 +19,10 @@ Future Release
     Thanks to the following people for contributing to this release:
     :user:`simha104`, :user:`tamargrey`, :user:`thehomebrewnerd`, :user:`tuethan1999`
 
+Breaking Changes
+++++++++++++++++
+    * The default logical type is now the ``Unknown`` type instead of the ``NaturalLanguage`` type. 
+      The global config ``natural_language_threshold`` has been renamed to ``categorical_threshold``.
 
 v0.4.2 Jun 23, 2021
 ===================