+++ docs/source/release_notes.rst
@@ -10,6 +10,8 @@ Future Release
     * Changes
         * Remove lines in ``LogicalType.transform`` that raise error if dtype conflicts (:pr:`1012`)
         * Add ``infer_datetime_format`` param to speed up ``to_datetime`` calls (:pr:`1016`)
+        * Add support for numpy array inputs to Woodwork (:pr:`1023`)
+        * Add support for pandas.api.extensions.ExtensionArray inputs to Woodwork (:pr:`1026`)
         * The default logical type is now the ``Unknown`` type instead of the ``NaturalLanguage`` type (:pr:`992`)
         * Add support for numpy inputs to Woodwork (:pr:`1023`)
     * Documentation Changes